Login, BottomNavBar y otros cambios

+++ app/src/main/java/com/ags/proyectofinal/ui/ListaPedidosFragment.kt
@@ -1,11 +1,13 @@
 package com.ags.proyectofinal.ui
 
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import androidx.fragment.app.Fragment
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -14,6 +16,8 @@ import com.ags.proyectofinal.application.ProyectoFinalApp
 import com.ags.proyectofinal.data.db.PedidoRepository
 import com.ags.proyectofinal.data.db.model.PedidoEntity
 import com.ags.proyectofinal.databinding.FragmentListaPedidosBinding
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.FirebaseUser
 import kotlinx.coroutines.launch
 import java.io.IOException
 
@@ -21,6 +25,10 @@ class ListaPedidosFragment : Fragment() {
 
     private var _binding : FragmentListaPedidosBinding ? = null
     private val binding get() = _binding!!
+
+    private lateinit var firebaseAuth: FirebaseAuth
+    private var user: FirebaseUser? = null
+    private var userId = ""
 
     private var pedidos : List<PedidoEntity> = emptyList()
     private lateinit var repository: PedidoRepository
@@ -39,6 +47,12 @@ class ListaPedidosFragment : Fragment() {
 
 
         repository = (activity?.application as ProyectoFinalApp).repository
+
+        firebaseAuth = FirebaseAuth.getInstance()
+        user = firebaseAuth.currentUser
+        if (user != null){
+            userId = user!!.uid
+        }
 
         pedidoAdapter = PedidoAdapter(){
             pedido,action -> onPedidoClicked(pedido, action)
@@ -72,17 +86,27 @@ class ListaPedidosFragment : Fragment() {
                     .commit()
             }
             "Delete" -> {
-                try{
-                    lifecycleScope.launch {
-                        repository.deletePedido(pedido)
-                        updateUI()
+                AlertDialog.Builder(requireContext())
+                    .setTitle(getString(R.string.eliminarPedido))
+                    .setMessage(getString(R.string.mensajeEliminarPedido))
+                    .setPositiveButton(getString(R.string.botonPositivo)){ _,_ ->
+                        try{
+                            lifecycleScope.launch {
+                                repository.deletePedido(pedido)
+                                updateUI()
+                                Toast.makeText(requireContext(), getString(R.string.eliminacionExitosa), Toast.LENGTH_SHORT).show()
 
+                            }
+                        }catch (e: IOException){
+                            e.printStackTrace()
+                            Toast.makeText(requireContext(), getString(R.string.eliminacionFallida), Toast.LENGTH_SHORT).show()
+                        }
+
+                    }.setNegativeButton(getString(R.string.botonNegativo)){ dialog,_ ->
+                        dialog.dismiss()
                     }
-                }catch (e: IOException){
-                    e.printStackTrace()
-                }
-
-
+                    .create()
+                    .show()
             }
 
             "Detail" ->{
@@ -96,12 +120,13 @@ class ListaPedidosFragment : Fragment() {
 
     private fun updateUI(){
         lifecycleScope.launch {
-            pedidos = repository.getAllPedidos()
+            //Toast.makeText(requireContext(), "ID: $userId", Toast.LENGTH_SHORT).show()
+            pedidos = repository.getAllPedidosByUser(userId)
 
             if (pedidos.isNotEmpty()) {
 
             }else{
-                Toast.makeText(requireContext(), getString(R.string.noRegistros), Toast.LENGTH_LONG).show()
+                //Toast.makeText(requireContext(), getString(R.string.noRegistros), Toast.LENGTH_LONG).show()
             }
             pedidoAdapter.updateList(pedidos)
         }
